2024/project-work/maticmind/requirements.txt
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
future==1.0.0
humanfriendly==10.0
idna==3.7
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.2.2
webex-bot==0.5.1
webexteamssdk==1.6.1
websockets==11.0.3
